switched accrual type

+++ internal/model/order.go
@@ -19,5 +19,5 @@ type Order struct {
 	UserID      uint       `db:"user_id" json:"-"`
 	Status      OrderState `db:"status" json:"status"`
 	UploadedAt  time.Time  `db:"uploaded_at" json:"uploaded_at"`
-	Accrual     int        `db:"accrual" json:"accrual,omitempty"`
+	Accrual     float64    `db:"accrual" json:"accrual,omitempty"`
 }
